svd: Put the XML declaration at the very top
According to the XML specification[1], the XML declaration (XMLDecl) is
allowed *only* as the very first production in the prolog, which in turn
can appear only as the very first production in the document. This means
it must come before all comments or the XML file is invalid. Java and
Python XML parsers fail[2] on documents like this one which violate that
part of the spec.

Move the declaration up so our SVD is valid XML.

[1] https://www.w3.org/TR/xml/#NT-prolog
[2] https://stackoverflow.com/a/19898942/6568778